backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
openai
requests
httpx
elevenlabs
psycopg2-binary
llama-cloud-services
